requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.1
fake-headers==1.0.2
html5lib==1.1
icalendar==4.1.0
idna==3.3
lxml==4.9.1
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.12
webencodings==0.5.1
